Fix requested changes

+++ cubic-server/protocol/ClientPackets.cpp
@@ -80,7 +80,7 @@ std::shared_ptr<std::vector<uint8_t>> protocol::createEntityAnimation(EntityAnim
 {
     std::vector<uint8_t> payload;
     serialize(payload, entityID, addVarInt,
-        (uint8_t)animId, addByte);
+        animId, addByte);
     auto packet = std::make_shared<std::vector<uint8_t>>();
     finalize(*packet, payload, (int32_t) ClientPacketID::EntityAnimation);
     return packet;
